+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 # Release Notes
 
+## OCD-1495
+* Add /resources/create API call to allow returning a CSV from JSON input
+
 ## OCD-1550
 * Improve /update API call speed. Selectively update only items that have changed rather than doing a replace on the entire listing.